requirements.txt
fastapi
uvicorn
requests
pymupdf
pdfplumber
sentence-transformers
faiss-cpu
numpy
streamlit
